Create tutorial.md
This tutorial shows the creation of a workflow capable of carrying out a variant calling analysis, using as input files coming from a study on the long-term evolution of an E. coli population (strain Ara -3). We'll carry out variant calling to see how the population has changed over time.

We want to understand how this population has changed compared to the original population, the REL606 strain of E. coli. Therefore, we will align each of our samples to the REL606 E. coli reference genome and see what differences exist in our reads compared to the genome.

The entire procedure is implemented to be performed comfortably with the interactive tools present in the Galaxy ToolShed.